Refactor attempt
This would simplify the code even more
